requirements.txt
pyspark # library for the apache spark
psycopg2-binary #DB conexion
sqlalchemy #DB ORM
pandas #Data Manipulation